# load-testing-swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Anecole Backend API - Load Testing",
    "description": "API спецификация для нагрузочного тестирования образовательной платформы Anecole",
    "version": "1.0.0",
    "contact": {
      "name": "Anecole Development Team",
      "email": "dev@anecole.com"
    }
  },
  "servers": [
    {
      "url": "http://localhost:8080/api",
      "description": "Local Development Server"
    },
    {
      "url": "http://10.10.3.4:8080/api",
      "description": "Production Server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "Lesson": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "example": 1
          },
          "title": {
            "type": "string",
            "example": "Урок 1: Приветствие"
          },
          "description": {
            "type": "string",
            "example": "Изучаем основные приветствия на английском языке"
          },
          "orderNumber": {
            "type": "integer",
            "example": 1
          },
          "active": {
            "type": "boolean",
            "example": true
          }
        }
      },
      "Word": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "example": 1
          },
          "word": {
            "type": "string",
            "example": "hello"
          },
          "translation": {
            "type": "string",
            "example": "привет"
          },
          "example": {
            "type": "string",
            "example": "Hello, how are you?"
          },
          "pronunciation": {
            "type": "string",
            "example": "həˈloʊ"
          }
        }
      },
      "StudentProgress": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "example": 1
          },
          "score": {
            "type": "integer",
            "example": 85
          },
          "completed": {
            "type": "boolean",
            "example": true
          },
          "word": {
            "$ref": "#/components/schemas/Word"
          }
        }
      }
    }
  },
  "paths": {
...
    "/lessons": {
      "get": {
        "tags": ["Уроки"],
        "summary": "Получить все уроки",
        "description": "Получение списка всех доступных уроков",
        "operationId": "getAllLessons",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Список уроков",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Lesson"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/lessons/{id}": {
      "get": {
        "tags": ["Уроки"],
        "summary": "Получить урок по ID",
        "description": "Получение информации о конкретном уроке",
        "operationId": "getLessonById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "example": 1
          }
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Информация об уроке",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Lesson"
                }
              }
            }
          },
          "404": {
            "description": "Урок не найден"
          }
        }
      }
    },
...
    "/students/{id}/progress": {
      "get": {
        "tags": ["Прогресс"],
        "summary": "Получить прогресс студента",
        "description": "Получение прогресса обучения студента",
        "operationId": "getStudentProgress",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "example": 1
          }
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Прогресс студента",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/StudentProgress"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/stats/dashboard": {
      "get": {
        "tags": ["Статистика"],
        "summary": "Получить статистику дашборда",
        "description": "Получение общей статистики для дашборда",
        "operationId": "getDashboardStats",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Статистика дашборда",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "totalLessons": {
                      "type": "integer",
                      "example": 11
                    },
                    "totalWords": {
                      "type": "integer",
                      "example": 32
                    },
                    "activeLessons": {
                      "type": "integer",
                      "example": 10
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/stats/student/{studentId}": {
      "get": {
        "tags": ["Статистика"],
        "summary": "Получить статистику студента",
        "description": "Получение статистики обучения конкретного студента",
        "operationId": "getStudentStats",
        "parameters": [
          {
            "name": "studentId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "example": 1
          }
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Статистика студента",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "completedWords": {
                      "type": "integer",
                      "example": 15
                    },
                    "averageScore": {
                      "type": "number",
                      "example": 87.5
                    },
                    "totalLessons": {
                      "type": "integer",
                      "example": 8
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/search/lessons": {
      "get": {
        "tags": ["Поиск"],
        "summary": "Поиск уроков",
        "description": "Поиск уроков по названию или описанию",
        "operationId": "searchLessons",
        "parameters": [
          {
            "name": "query",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "приветствие"
          }
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Результаты поиска",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Lesson"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/search/words": {
      "get": {
        "tags": ["Поиск"],
        "summary": "Поиск слов",
        "description": "Поиск слов по названию или переводу",
        "operationId": "searchWords",
        "parameters": [
          {
            "name": "query",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "hello"
          }
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Результаты поиска",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Word"
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
...
} 